Update Documentation for Auto Complete
