--- docs/src/app/components/raw-code/auto-complete-code.txt
@@ -1,48 +1,0 @@
-<AutoComplete
-  dataSource={this.state.input1}
-  onUpdateInput={this.handleUpdateInput1}
-  onNewRequest={this.handleNewRequest}
-/>
-<br/>
-<AutoComplete
-  hintText="hint"
-  dataSource={this.state.input2}
-  onUpdateInput={this.handleUpdateInput2}
-  onNewRequest={this.handleNewRequest}
-/>
-<br/>
-<AutoComplete
-  fullWidth={true}
-  searchText="*****"
-  errorText="error message"
-  dataSource={this.state.input3}
-  onUpdateInput={this.handleUpdateInput3}
-  onNewRequest={this.handleNewRequest}
-/>
-<br/>
-<AutoComplete
-  hintText="text-value data"
-  filter={AutoComplete.noFilter}
-  dataSource={dataSource}
-  onNewRequest={this.handleNewRequest}
-/>
-<br/>
-<AutoComplete
-  floatingLabelText="fuzzy search"
-  filter={AutoComplete.fuzzyFilter}
-  dataSource={fruit}
-/>
-<br/>
-<AutoComplete
-  floatingLabelText="case insensitive, colors"
-  filter={AutoComplete.caseInsensitiveFilter}
-  dataSource={colors}
-/>
-<br/>
-<AutoComplete
-  floatingLabelText="showAllItems"
-  filter={AutoComplete.noFilter}
-  triggerUpdateOnFocus={true}
-  animated={false}
-  dataSource={['12345', '23456', '34567']}
-/>
